pyproject.toml
[project]
name = "chat-seguro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
